Refactoring - Update dependencies - Extract code downloading subtitles to an interface to ease the integration of new providers
