+++ src/jonag/Episodes/Exception/InvalidProviderException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace jonag\Episodes\Exception;
+
+
+class InvalidProviderException extends \InvalidArgumentException
+{
+
+}
