+++ src/jonag/Episodes/Provider/ProviderInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace jonag\Episodes\Provider;
+
+use Symfony\Component\Console\Style\OutputStyle;
+
+interface ProviderInterface
+{
+    /**
+     * @param \Symfony\Component\Console\Style\OutputStyle $io
+     * @param \SplFileInfo                                 $fileInfo
+     * @return null|string
+     */
+    public function findSubtitleForFile(OutputStyle $io, \SplFileInfo $fileInfo);
+}
